Consistently use ID3-based endpoints
The subsonic API differentiates between organizing the music by file
structure vs. ID3 tags.

Previously, the albums listed on the main page would be retrieved from
the `getAlbumList` endpoint (file structure based), and their `id`'s
would then be passed to `getAlbum` (ID3-based) to retrieve songs and
other information. Since the `id` from `getAlbumList` referred to a
folder ID and `getAlbum` was expected an album id, this would return
bad/no data.

This commit changes the album listing endpoint that's used to the
ID3-based `getAlbumList2` endpoint. This ensures that the `id` property
of the albums always refers to an ID3-based album, not a folder.
Similarly, the `getAlbumInfo` endpoint was changed to `getAlbumInfo2`.

This commit also changes all references of an album's `title` attribute
to `name` since `title` is only returned when using file structure
endpoints.